Updated flavour text
I added a few small brief paragraphs that give an overview of what the app does, but I want the paragraphs to go along horizontally rather than vertically and I will need to learn how to do that (unless somebody gets around to it before me). Also, I think it would be great if we could use some small graphic to go with each 'block'.